refactor: rename reducer to match corresponding action
